+++ src/structures/paging/mapper/mod.rs
@@ -1,10 +1,10 @@
 //! Abstractions for reading and modifying the mapping of pages.
 
-pub use self::mapped_page_table::{MappedPageTable, PhysToVirt};
+pub use self::mapped_page_table::{MappedPageTable, PageTableFrameMapping};
 #[cfg(target_pointer_width = "64")]
 pub use self::offset_page_table::OffsetPageTable;
 #[cfg(feature = "instructions")]
-pub use self::recursive_page_table::RecursivePageTable;
+pub use self::recursive_page_table::{InvalidPageTable, RecursivePageTable};
 
 use crate::structures::paging::{
     frame_alloc::FrameAllocator, page_table::PageTableFlags, Page, PageSize, PhysFrame, Size1GiB,
@@ -17,9 +17,13 @@ mod offset_page_table;
 #[cfg(feature = "instructions")]
 mod recursive_page_table;
 
-/// This trait defines page table operations that work for all page sizes of the x86_64
-/// architecture.
-pub trait MapperAllSizes: Mapper<Size4KiB> + Mapper<Size2MiB> + Mapper<Size1GiB> {
+/// An empty convencience trait that requires the `Mapper` trait for all page sizes.
+pub trait MapperAllSizes: Mapper<Size4KiB> + Mapper<Size2MiB> + Mapper<Size1GiB> {}
+
+impl<T> MapperAllSizes for T where T: Mapper<Size4KiB> + Mapper<Size2MiB> + Mapper<Size1GiB> {}
+
+/// Provides methods for translating virtual addresses.
+pub trait Translate {
     /// Return the frame that the given virtual address is mapped to and the offset within that
     /// frame.
     ///
@@ -34,49 +38,74 @@ pub trait MapperAllSizes: Mapper<Size4KiB> + Mapper<Size2MiB> + Mapper<Size1GiB>
     /// Returns `None` if there is no valid mapping for the given address.
     ///
     /// This is a convenience method. For more information about a mapping see the
-    /// [`translate`](MapperAllSizes::translate) method.
+    /// [`translate`](Translate::translate) method.
     #[inline]
     fn translate_addr(&self, addr: VirtAddr) -> Option<PhysAddr> {
         match self.translate(addr) {
-            TranslateResult::PageNotMapped | TranslateResult::InvalidFrameAddress(_) => None,
-            TranslateResult::Frame4KiB { frame, offset } => Some(frame.start_address() + offset),
-            TranslateResult::Frame2MiB { frame, offset } => Some(frame.start_address() + offset),
-            TranslateResult::Frame1GiB { frame, offset } => Some(frame.start_address() + offset),
+            TranslateResult::NotMapped | TranslateResult::InvalidFrameAddress(_) => None,
+            TranslateResult::Mapped { frame, offset, .. } => Some(frame.start_address() + offset),
         }
     }
 }
 
-/// The return value of the [`MapperAllSizes::translate`] function.
+/// The return value of the [`Translate::translate`] function.
 ///
 /// If the given address has a valid mapping, a `Frame4KiB`, `Frame2MiB`, or `Frame1GiB` variant
 /// is returned, depending on the size of the mapped page. The remaining variants indicate errors.
 #[derive(Debug)]
 pub enum TranslateResult {
-    /// The page is mapped to a physical frame of size 4KiB.
-    Frame4KiB {
+    /// The virtual address is mapped to a physical frame.
+    Mapped {
         /// The mapped frame.
-        frame: PhysFrame<Size4KiB>,
+        frame: MappedFrame,
         /// The offset whithin the mapped frame.
         offset: u64,
+        /// The entry flags in the lowest-level page table.
+        ///
+        /// Flags of higher-level page table entries are not included here, but they can still
+        /// affect the effective flags for an address, for example when the WRITABLE flag is not
+        /// set for a level 3 entry.
+        flags: PageTableFlags,
     },
-    /// The page is mapped to a physical frame of size 2MiB.
-    Frame2MiB {
-        /// The mapped frame.
-        frame: PhysFrame<Size2MiB>,
-        /// The offset whithin the mapped frame.
-        offset: u64,
-    },
-    /// The page is mapped to a physical frame of size 2MiB.
-    Frame1GiB {
-        /// The mapped frame.
-        frame: PhysFrame<Size1GiB>,
-        /// The offset whithin the mapped frame.
-        offset: u64,
-    },
-    /// The given page is not mapped to a physical frame.
-    PageNotMapped,
-    /// The page table entry for the given page points to an invalid physical address.
+    /// The given virtual address is not mapped to a physical frame.
+    NotMapped,
+    /// The page table entry for the given virtual address points to an invalid physical address.
     InvalidFrameAddress(PhysAddr),
+}
+
+/// Represents a physical frame mapped in a page table.
+#[derive(Debug)]
+pub enum MappedFrame {
+    /// The virtual address is mapped to a 4KiB frame.
+    Size4KiB(PhysFrame<Size4KiB>),
+    /// The virtual address is mapped to a "large" 2MiB frame.
+    Size2MiB(PhysFrame<Size2MiB>),
+    /// The virtual address is mapped to a "huge" 1GiB frame.
+    Size1GiB(PhysFrame<Size1GiB>),
+}
+
+impl MappedFrame {
+    const_fn! {
+        /// Returns the start address of the frame.
+        pub fn start_address(&self) -> PhysAddr {
+            match self {
+                MappedFrame::Size4KiB(frame) => frame.start_address(),
+                MappedFrame::Size2MiB(frame) => frame.start_address(),
+                MappedFrame::Size1GiB(frame) => frame.start_address(),
+            }
+        }
+    }
+
+    const_fn! {
+        /// Returns the size the frame (4KB, 2MB or 1GB).
+        pub fn size(&self) -> u64 {
+            match self {
+                MappedFrame::Size4KiB(frame) => frame.size(),
+                MappedFrame::Size2MiB(frame) => frame.size(),
+                MappedFrame::Size1GiB(frame) => frame.size(),
+            }
+        }
+    }
 }
 
 /// A trait for common page table operations on pages of size `S`.
@@ -158,7 +187,7 @@ pub trait Mapper<S: PageSize> {
     ) -> Result<MapperFlush<S>, MapToError<S>>
     where
         Self: Sized,
-        A: FrameAllocator<Size4KiB>,
+        A: FrameAllocator<Size4KiB> + ?Sized,
     {
         let parent_table_flags = flags
             & (PageTableFlags::PRESENT
@@ -248,7 +277,7 @@ pub trait Mapper<S: PageSize> {
     ) -> Result<MapperFlush<S>, MapToError<S>>
     where
         Self: Sized,
-        A: FrameAllocator<Size4KiB>;
+        A: FrameAllocator<Size4KiB> + ?Sized;
 
     /// Removes a mapping from the page table and returns the frame that used to be mapped.
     ///
@@ -336,7 +365,7 @@ pub trait Mapper<S: PageSize> {
     ) -> Result<MapperFlush<S>, MapToError<S>>
     where
         Self: Sized,
-        A: FrameAllocator<Size4KiB>,
+        A: FrameAllocator<Size4KiB> + ?Sized,
         S: PageSize,
         Self: Mapper<S>,
     {
@@ -448,4 +477,4 @@ pub enum TranslateError {
     InvalidFrameAddress(PhysAddr),
 }
 
-static _ASSERT_OBJECT_SAFE: Option<&(dyn MapperAllSizes + Sync)> = None;
+static _ASSERT_OBJECT_SAFE: Option<&(dyn Translate + Sync)> = None;