+++ src/structures/paging/mod.rs
@@ -3,8 +3,6 @@
 //! Page tables translate virtual memory “pages” to physical memory “frames”.
 
 pub use self::frame::PhysFrame;
-#[allow(deprecated)]
-pub use self::frame_alloc::UnusedPhysFrame;
 pub use self::frame_alloc::{FrameAllocator, FrameDeallocator};
 #[doc(no_inline)]
 pub use self::mapper::MappedPageTable;
@@ -14,7 +12,7 @@ pub use self::mapper::OffsetPageTable;
 #[cfg(feature = "instructions")]
 #[doc(no_inline)]
 pub use self::mapper::RecursivePageTable;
-pub use self::mapper::{Mapper, MapperAllSizes};
+pub use self::mapper::{Mapper, Translate};
 pub use self::page::{Page, PageSize, Size1GiB, Size2MiB, Size4KiB};
 pub use self::page_table::{PageOffset, PageTable, PageTableFlags, PageTableIndex};